+++ app/components/bulma/content_component.rb
@@ -2,8 +2,15 @@
 
 module Bulma
   class ContentComponent < Component
+    def initialize(size: nil)
+      @size = size
+    end
+
     def call
-      content_tag :div, content, class: "content"
+      content_tag :div, content, class: class_names(
+        "content",
+        "is-#{@size}" => @size.present?
+      )
     end
   end
 end